+++ src/thedaemon/__main__.py
@@ -9,6 +9,13 @@ from .server import bind_and_listen
 def main() -> None:
     parser = argparse.ArgumentParser()
     parser.set_defaults(command=lambda args: parser.print_help())
+    parser.add_argument(
+        "-P",
+        "--port",
+        default=21365,
+        help="The daemon port to use. (default: %(default)s)",
+        type=int,
+    )
     subcmd = parser.add_subparsers()
 
     start = subcmd.add_parser("start", help="Start the daemon process.")
@@ -32,33 +39,33 @@ def main() -> None:
 
 # Commands
 def start_server(args) -> None:
-    check_server_not_running()
+    check_server_not_running(port=args.port)
 
     if args.daemon:
-        daemonize(win_args=["-m", "thedaemon", "start"])
+        daemonize(win_args=["-m", "thedaemon", "-P", str(args.port), "start"])
 
-    bind_and_listen()
+    bind_and_listen(port=args.port)
 
 
 def ping_server(args) -> None:
-    with connect_to_server() as client:
+    with connect_to_server(port=args.port) as client:
         client.sendall(b"ping\n")
         print(client.recv(1024).decode().rstrip())
 
 
 def stop_server(args) -> None:
-    with connect_to_server() as client:
+    with connect_to_server(port=args.port) as client:
         client.sendall(b"stop\n")
         print(client.recv(1024).decode().rstrip())
 
 
 # Utilities
-def check_server_not_running() -> None:
+def check_server_not_running(*, port: int) -> None:
     # Normally if the server is already running, binding will raise OSError.
     # However, let's try connecting to it before daemonization so we can give
     # a useful message to the user.
     try:
-        with connect_to_server() as client:
+        with connect_to_server(port=port) as client:
             pass
     except SystemExit:
         pass