+++ src/thedaemon/client.py
@@ -1,10 +1,10 @@
 import socket
 import sys
 
-from .constants import ADDRESS, PORT
+from .constants import ADDRESS
 
 
-def connect_to_server() -> socket.socket:
+def connect_to_server(*, port: int) -> socket.socket:
     client = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
 
     if sys.platform == "win32":
@@ -13,7 +13,7 @@ def connect_to_server() -> socket.socket:
         client.settimeout(0.1)
 
     try:
-        client.connect((ADDRESS, PORT))
+        client.connect((ADDRESS, port))
     except (ConnectionRefusedError, TimeoutError):
         client.close()
         sys.exit("Failed to connect to daemon; did you start it?")